requirements.txt
# requirements.txt
streamlit==1.28.0
neo4j==5.14.0
numpy==1.24.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.0